Create MassTransitModel
