+++ MassTransitModel/Enums/QueueEnum.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MassTransitService.Enums
+namespace MassTransitCommon.Enums
 {
     public enum QueueEnum
     {
