+++ MassTransitModel/MassTransitCommon.csproj
