+++ MassTransitModel/Model/ReportModel.cs
@@ -5,7 +5,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace MassTransitService.Model
+namespace MassTransitCommon.Model
 {
     public class ReportModel
     {
