+++ MicroservisPhoneBook.sln
@@ -15,7 +15,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WorkerService", "WorkerServ
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Common", "Common", "{ECC6844B-E982-4D35-893D-7031E63F873E}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MassTransitService", "MassTransitModel\MassTransitService.csproj", "{2DC7CF0C-7A6F-4359-9860-CB2354AF7588}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MassTransitCommon", "MassTransitModel\MassTransitCommon.csproj", "{2DC7CF0C-7A6F-4359-9860-CB2354AF7588}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -47,7 +47,6 @@ Global
 		{A55EAE8D-4CF9-44F4-BC85-330FD7CE797A} = {A7EDBAA8-F658-40DC-9A92-9086498668BA}
 		{5CC0844C-9BBC-4DE9-9A0A-C7EE4FF3B852} = {A7EDBAA8-F658-40DC-9A92-9086498668BA}
 		{B658EB4E-07AE-4181-A31A-C774072D3C13} = {5A5EE81E-3EB2-4F15-827E-0066E6CDBA70}
-		{2DC7CF0C-7A6F-4359-9860-CB2354AF7588} = {ECC6844B-E982-4D35-893D-7031E63F873E}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {3EEBA8B0-7244-440F-A274-32CFE4E9A68C}
