+++ PhoneBookService/Domain/DTO's/Report/ReportDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace PhoneBookService.Domain.DTO_s.Report
+{
+    public class ReportDTO
+    {
+        public int MyProperty { get; set; }
+    }
+}
